Split several Process files for Linux
System.Diagnostics.Process.dll's implementation on Linux uses some standard Unix functionality and some Linux-specific functionality.  This commit just splits the .Linux.cs files into .Unix.cs and .Linux.cs files, with everything expected to work on any Unix system moved into the .Unix.cs files.  This is one step towards implementing this assembly for OS X as well (at which point we can move things around further as needed).

No code was actually changed in this commit, just moved between files.
