+++ src/System.Diagnostics.Process/src/System/Diagnostics/ProcessManager.Linux.cs
@@ -1,7 +1,6 @@
 // Copyright (c) Microsoft. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-using Microsoft.Win32.SafeHandles;
 using System.Collections.Generic;
 using System.Globalization;
 using System.IO;
@@ -16,57 +15,6 @@ namespace System.Diagnostics
         public static bool IsProcessRunning(int processId)
         {
             return File.Exists(Interop.procfs.GetStatFilePathForProcess(processId));
-        }
-
-        /// <summary>Gets whether the process with the specified ID on the specified machine is currently running.</summary>
-        /// <param name="processId">The process ID.</param>
-        /// <param name="machineName">The machine name.</param>
-        /// <returns>true if the process is running; otherwise, false.</returns>
-        public static bool IsProcessRunning(int processId, string machineName)
-        {
-            ThrowIfRemoteMachine(machineName);
-            return IsProcessRunning(processId);
-        }
-
-        /// <summary>Gets the ProcessInfo for the specified process ID on the specified machine.</summary>
-        /// <param name="processId">The process ID.</param>
-        /// <param name="machineName">The machine name.</param>
-        /// <returns>The ProcessInfo for the process if it could be found; otherwise, null.</returns>
-        public static ProcessInfo GetProcessInfo(int processId, string machineName)
-        {
-            ThrowIfRemoteMachine(machineName);
-            return CreateProcessInfo(processId);
-        }
-
-        /// <summary>Gets process infos for each process on the specified machine.</summary>
-        /// <param name="machineName">The target machine.</param>
-        /// <returns>An array of process infos, one per found process.</returns>
-        public static ProcessInfo[] GetProcessInfos(string machineName)
-        {
-            ThrowIfRemoteMachine(machineName);
-            int[] procIds = GetProcessIds(machineName);
-
-            // Iterate through all process IDs to load information about each process
-            var processes = new List<ProcessInfo>(procIds.Length);
-            foreach (int pid in procIds)
-            {
-                ProcessInfo pi = CreateProcessInfo(pid);
-                if (pi != null)
-                {
-                    processes.Add(pi);
-                }
-            }
-
-            return processes.ToArray();
-        }
-
-        /// <summary>Gets the IDs of all processes on the specified machine.</summary>
-        /// <param name="machineName">The machine to examine.</param>
-        /// <returns>An array of process IDs from the specified machine.</returns>
-        public static int[] GetProcessIds(string machineName)
-        {
-            ThrowIfRemoteMachine(machineName);
-            return GetProcessIds();
         }
 
         /// <summary>Gets the IDs of all processes on the current machine.</summary>
@@ -88,46 +36,9 @@ namespace System.Diagnostics
             return pids.ToArray();
         }
 
-        /// <summary>Gets the ID of a process from a handle to the process.</summary>
-        /// <param name="processHandle">The handle.</param>
-        /// <returns>The process ID.</returns>
-        public static int GetProcessIdFromHandle(SafeProcessHandle processHandle)
-        {
-            return (int)processHandle.DangerousGetHandle(); // not actually dangerous; just wraps a process ID
-        }
-
-        /// <summary>Gets an array of module infos for the specified process.</summary>
-        /// <param name="processId">The ID of the process whose modules should be enumerated.</param>
-        /// <returns>The array of modules.</returns>
-        public static ModuleInfo[] GetModuleInfos(int processId)
-        {
-            // Not currently supported, but we can simply return an empty array rather than throwing.
-            // Could potentially be done via /proc/pid/maps and some heuristics to determine
-            // which entries correspond to modules.
-            return Array.Empty<ModuleInfo>();
-        }
-
-        /// <summary>Gets whether the named machine is remote or local.</summary>
-        /// <param name="machineName">The machine name.</param>
-        /// <returns>true if the machine is remote; false if it's local.</returns>
-        public static bool IsRemoteMachine(string machineName)
-        {
-            return 
-                machineName != "." && 
-                machineName != Interop.libc.gethostname();
-        }
-
         // -----------------------------
         // ---- PAL layer ends here ----
         // -----------------------------
-
-        private static void ThrowIfRemoteMachine(string machineName)
-        {
-            if (IsRemoteMachine(machineName))
-            {
-                throw new PlatformNotSupportedException();
-            }
-        }
 
         private static ProcessInfo CreateProcessInfo(int pid)
         {
